unify resizing sections
