+++ editor/src/components/inspector/fill-hug-fixed-control.tsx
@@ -5,8 +5,8 @@ import React from 'react'
 import { createSelector } from 'reselect'
 import { optionalMap } from '../../core/shared/optional-utils'
 import { intersection } from '../../core/shared/set-utils'
-import { assertNever, NO_OP } from '../../core/shared/utils'
-import { NumberInput, PopupList, UtopiaTheme } from '../../uuiui'
+import { assertNever } from '../../core/shared/utils'
+import { NumberInput, PopupList, useColorTheme, UtopiaTheme } from '../../uuiui'
 import type { SelectOption } from '../../uuiui-deps'
 import { getControlStyles, InspectorRowHoverCSS } from '../../uuiui-deps'
 import { useDispatch } from '../editor/store/dispatch-context'
@@ -36,13 +36,21 @@ import type { MetadataSubstate } from '../editor/store/store-hook-substore-types
 import type { ElementPath } from '../../core/shared/project-file-types'
 import { treatElementAsGroupLike } from '../canvas/canvas-strategies/strategies/group-helpers'
 import * as EP from '../../core/shared/element-path'
-import { FlexCol } from 'utopia-api'
+import * as PP from '../../core/shared/property-path'
 import { UIGridRow } from './widgets/ui-grid-row'
-import { PropertyLabel } from './widgets/property-label'
 import { useContextSelector } from 'use-context-selector'
 import { InspectorPropsContext, stylePropPathMappingFn } from './common/property-path-hooks'
-import { safeIndex } from '../../core/shared/array-utils'
+import { mapDropNulls, safeIndex, uniqBy } from '../../core/shared/array-utils'
 import { fixedSizeDimensionHandlingText } from '../text-editor/text-handling'
+import { unless, when } from '../../utils/react-conditionals'
+import type { LayoutPinnedProp } from '../../core/layout/layout-helpers-new'
+import type { AllElementProps } from '../editor/store/editor-state'
+import { jsExpressionValue, emptyComments } from '../../core/shared/element-template'
+import type { EditorDispatch, EditorAction } from '../editor/action-types'
+import { unsetProperty, setProp_UNSAFE } from '../editor/actions/action-creators'
+import { FlexCol } from 'utopia-api'
+import { PinControl } from './controls/pin-control'
+import type { FramePinsInfo } from './common/layout-property-path-hooks'
 
 export const FillFixedHugControlId = (segment: 'width' | 'height'): string =>
   `hug-fixed-fill-${segment}`
@@ -124,6 +132,7 @@ export const FillHugFixedControl = React.memo<FillHugFixedControlProps>((props) 
   const allElementPropsRef = useRefEditorState((store) => store.editor.allElementProps)
 
   const elementOrParentGroupRef = useRefEditorState(anySelectedElementGroupOrChildOfGroup)
+  const groupChildrenSelectedRef = useRefEditorState(allElementsAreGroupChildren)
 
   const widthCurrentValue = useEditorState(
     Substores.metadata,
@@ -375,6 +384,87 @@ export const FillHugFixedControl = React.memo<FillHugFixedControlProps>((props) 
     ],
   )
 
+  const editorRef = useRefEditorState((store) => ({
+    selectedViews: store.editor.selectedViews,
+    allElementProps: store.editor.allElementProps,
+  }))
+
+  const groupChildConstraints = useEditorState(
+    Substores.metadata,
+    (store) => ({
+      width: checkGroupChildConstraint(
+        'width',
+        store.editor.selectedViews,
+        store.editor.allElementProps,
+      ),
+      height: checkGroupChildConstraint(
+        'height',
+        store.editor.selectedViews,
+        store.editor.allElementProps,
+      ),
+      top: checkGroupChildConstraint(
+        'top',
+        store.editor.selectedViews,
+        store.editor.allElementProps,
+      ),
+      left: checkGroupChildConstraint(
+        'left',
+        store.editor.selectedViews,
+        store.editor.allElementProps,
+      ),
+      bottom: checkGroupChildConstraint(
+        'bottom',
+        store.editor.selectedViews,
+        store.editor.allElementProps,
+      ),
+      right: checkGroupChildConstraint(
+        'right',
+        store.editor.selectedViews,
+        store.editor.allElementProps,
+      ),
+    }),
+    'FillHugFixedControl constraints',
+  )
+
+  const handleGroupConstraintPinMouseDown = React.useCallback(
+    (dimension: LayoutPinnedProp) => {
+      return setGroupChildConstraint(
+        dispatch,
+        'toggle',
+        editorRef.current.selectedViews,
+        editorRef.current.allElementProps,
+        dimension,
+      )
+    },
+    [dispatch, editorRef],
+  )
+
+  const framePoints: FramePinsInfo = React.useMemo(() => {
+    const ignore = { isPrimaryPosition: false, isRelativePosition: false }
+    return {
+      left: {
+        isPrimaryPosition: groupChildConstraints.left !== 'not-constrained',
+        isRelativePosition: false,
+      },
+      top: {
+        isPrimaryPosition: groupChildConstraints.top !== 'not-constrained',
+        isRelativePosition: false,
+      },
+      bottom: {
+        isPrimaryPosition: groupChildConstraints.bottom !== 'not-constrained',
+        isRelativePosition: false,
+      },
+      right: {
+        isPrimaryPosition: groupChildConstraints.right !== 'not-constrained',
+        isRelativePosition: false,
+      },
+      width: ignore,
+      height: ignore,
+      centerX: ignore,
+      centerY: ignore,
+    }
+  }, [groupChildConstraints])
+
   if (options == null) {
     return null
   }
@@ -384,59 +474,178 @@ export const FillHugFixedControl = React.memo<FillHugFixedControlProps>((props) 
 
   return (
     <FlexCol css={{ paddingBottom: UtopiaTheme.layout.rowHorizontalPadding }}>
-      <UIGridRow padded variant='|20px|<--1fr--><--1fr-->' css={InspectorRowHoverCSS}>
-        <PropertyLabel target={widthProp}>W</PropertyLabel>
-        <PopupList
-          value={optionalMap(selectOption, widthCurrentValue.fixedHugFill?.type) ?? undefined}
-          options={options}
-          onSubmitValue={onSubmitFixedFillHugType['width']}
-          controlStyles={widthControlStyles}
-        />
-        <NumberInput
-          id={FillFixedHugControlId('width')}
-          testId={FillFixedHugControlId('width')}
-          value={widthValue}
-          onSubmitValue={onAdjustWidth}
-          onTransientSubmitValue={onAdjustWidth}
-          onForcedSubmitValue={onAdjustWidth}
-          controlStatus={widthInputControlStatus}
-          numberType={pickNumberType(widthCurrentValue.fixedHugFill)}
-          incrementControls={true}
-          stepSize={1}
-          minimum={0}
-          maximum={Infinity}
-          defaultUnitToHide={null}
-          focusOnMount={false}
-        />
-      </UIGridRow>
-      <UIGridRow padded variant='|20px|<--1fr--><--1fr-->' css={InspectorRowHoverCSS}>
-        <PropertyLabel target={heightProp}>H</PropertyLabel>
-        <PopupList
-          value={optionalMap(selectOption, heightCurrentValue.fixedHugFill?.type) ?? undefined}
-          options={options}
-          onSubmitValue={onSubmitFixedFillHugType['height']}
-          controlStyles={heightControlStyles}
-        />
-        <NumberInput
-          id={FillFixedHugControlId('height')}
-          testId={FillFixedHugControlId('height')}
-          value={heightValue}
-          onSubmitValue={onAdjustHeight}
-          onTransientSubmitValue={onAdjustHeight}
-          onForcedSubmitValue={onAdjustHeight}
-          controlStatus={heightInputControlStatus}
-          numberType={pickNumberType(heightCurrentValue.fixedHugFill)}
-          incrementControls={true}
-          stepSize={1}
-          minimum={0}
-          maximum={Infinity}
-          defaultUnitToHide={null}
-          focusOnMount={false}
-        />
+      <UIGridRow padded variant='<-auto-><----------1fr--------->'>
+        {when(
+          groupChildrenSelectedRef.current,
+          <PinControl
+            handlePinMouseDown={handleGroupConstraintPinMouseDown}
+            framePoints={framePoints}
+            controlStatus='simple'
+            exclude={{ center: true }}
+            name='group-child-controls'
+          />,
+        )}
+        <div>
+          <UIGridRow
+            padded={false}
+            variant={
+              groupChildrenSelectedRef.current
+                ? '|--67px--|<--------1fr-------->'
+                : '<--1fr--><--1fr-->'
+            }
+            css={InspectorRowHoverCSS}
+          >
+            <NumberInput
+              labelInner={'W'}
+              id={FillFixedHugControlId('width')}
+              testId={FillFixedHugControlId('width')}
+              value={widthValue}
+              onSubmitValue={onAdjustWidth}
+              onTransientSubmitValue={onAdjustWidth}
+              onForcedSubmitValue={onAdjustWidth}
+              controlStatus={widthInputControlStatus}
+              numberType={pickNumberType(widthCurrentValue.fixedHugFill)}
+              incrementControls={true}
+              stepSize={1}
+              minimum={0}
+              maximum={Infinity}
+              defaultUnitToHide={null}
+              focusOnMount={false}
+            />
+            {unless(
+              groupChildrenSelectedRef.current,
+              <PopupList
+                value={optionalMap(selectOption, widthCurrentValue.fixedHugFill?.type) ?? undefined}
+                options={options}
+                onSubmitValue={onSubmitFixedFillHugType['width']}
+                controlStyles={widthControlStyles}
+              />,
+            )}
+            {when(
+              groupChildrenSelectedRef.current,
+              <GroupConstraintSelect dimension='width' type={groupChildConstraints.width} />,
+            )}
+          </UIGridRow>
+          <UIGridRow
+            padded={false}
+            variant={
+              groupChildrenSelectedRef.current
+                ? '|--67px--|<--------1fr-------->'
+                : '<--1fr--><--1fr-->'
+            }
+            css={InspectorRowHoverCSS}
+          >
+            <NumberInput
+              labelInner={'H'}
+              id={FillFixedHugControlId('height')}
+              testId={FillFixedHugControlId('height')}
+              value={heightValue}
+              onSubmitValue={onAdjustHeight}
+              onTransientSubmitValue={onAdjustHeight}
+              onForcedSubmitValue={onAdjustHeight}
+              controlStatus={heightInputControlStatus}
+              numberType={pickNumberType(heightCurrentValue.fixedHugFill)}
+              incrementControls={true}
+              stepSize={1}
+              minimum={0}
+              maximum={Infinity}
+              defaultUnitToHide={null}
+              focusOnMount={false}
+            />
+            {unless(
+              groupChildrenSelectedRef.current,
+              <PopupList
+                value={
+                  optionalMap(selectOption, heightCurrentValue.fixedHugFill?.type) ?? undefined
+                }
+                options={options}
+                onSubmitValue={onSubmitFixedFillHugType['height']}
+                controlStyles={heightControlStyles}
+              />,
+            )}
+            {when(
+              groupChildrenSelectedRef.current,
+              <GroupConstraintSelect dimension='height' type={groupChildConstraints.height} />,
+            )}
+          </UIGridRow>
+        </div>
       </UIGridRow>
     </FlexCol>
   )
 })
+
+const GroupConstraintSelect = React.memo(
+  ({
+    dimension,
+    type,
+  }: {
+    dimension: LayoutPinnedProp
+    type: GroupChildConstraintOptionType | 'mixed'
+  }) => {
+    const dispatch = useDispatch()
+    const colorTheme = useColorTheme()
+
+    const editorRef = useRefEditorState((store) => ({
+      selectedViews: store.editor.selectedViews,
+      allElementProps: store.editor.allElementProps,
+    }))
+
+    function optionFromType(value: GroupChildConstraintOptionType | 'mixed') {
+      return value === 'constrained'
+        ? groupChildConstraintOptionValues.constrained
+        : value === 'not-constrained'
+        ? groupChildConstraintOptionValues.notConstrained
+        : groupChildConstraintOptionValues.mixed
+    }
+
+    const listValue = React.useMemo(() => {
+      return optionFromType(type)
+    }, [type])
+
+    const mainColor = React.useMemo(() => {
+      switch (type) {
+        case 'constrained':
+          return colorTheme.brandNeonPink.value
+        case 'mixed':
+          return colorTheme.primary.value
+        case 'not-constrained':
+          return colorTheme.subduedForeground.value
+        default:
+          assertNever(type)
+      }
+    }, [type, colorTheme])
+
+    const onSubmitValue = React.useCallback(
+      (option: SelectOption) => {
+        return setGroupChildConstraint(
+          dispatch,
+          option.value,
+          editorRef.current.selectedViews,
+          editorRef.current.allElementProps,
+          dimension,
+        )
+      },
+      [dispatch, editorRef, dimension],
+    )
+
+    return (
+      <PopupList
+        id={`group-child-resize-${dimension}`}
+        onSubmitValue={onSubmitValue}
+        value={listValue}
+        options={options}
+        style={{ position: 'relative' }}
+        containerMode={type === 'constrained' ? 'default' : 'showBorderOnHover'}
+        controlStyles={{
+          ...getControlStyles('simple'),
+          mainColor: mainColor,
+        }}
+      />
+    )
+  },
+)
+
+GroupConstraintSelect.displayName = 'GroupConstraintSelect'
 
 function strategyForChangingFillFixedHugType(
   fixedValue: number,
@@ -504,3 +713,107 @@ const anySelectedElementGroupOrChildOfGroup = createSelector(
     return selectedViews.some(elementOrAnyChildGroup)
   },
 )
+
+const allElementsAreGroupChildren = createSelector(
+  metadataSelector,
+  (store: MetadataSubstate) => store.editor.elementPathTree,
+  selectedViewsSelector,
+  (metadata, pathTrees, selectedViews): boolean => {
+    function elementOrAnyChildGroup(path: ElementPath) {
+      return treatElementAsGroupLike(metadata, EP.parentPath(path))
+    }
+    return selectedViews.every(elementOrAnyChildGroup)
+  },
+)
+
+function checkGroupChildConstraint(
+  dimension: LayoutPinnedProp,
+  selectedViews: ElementPath[],
+  allElementProps: AllElementProps,
+): GroupChildConstraintOptionType | 'mixed' {
+  const constrained = selectedViews.filter((path) =>
+    getSafeGroupChildConstraintsArray(allElementProps, path).includes(dimension),
+  ).length
+  return constrained === selectedViews.length
+    ? 'constrained'
+    : constrained === 0
+    ? 'not-constrained'
+    : 'mixed'
+}
+
+const groupChildConstraintOptionValues = {
+  constrained: groupChildConstraintOption('constrained'),
+  notConstrained: groupChildConstraintOption('not-constrained'),
+  mixed: { value: 'mixed', label: 'Mixed' },
+}
+
+const options: Array<SelectOption> = [
+  groupChildConstraintOptionValues.constrained,
+  groupChildConstraintOptionValues.notConstrained,
+]
+
+type GroupChildConstraintOptionType = 'constrained' | 'not-constrained'
+
+function groupChildConstraintOption(type: GroupChildConstraintOptionType): SelectOption {
+  switch (type) {
+    case 'constrained':
+      return { value: 'constrained', label: 'Constrained' }
+    case 'not-constrained':
+      return { value: 'not-constrained', label: 'Unconstrained' }
+    default:
+      assertNever(type)
+  }
+}
+
+function getSafeGroupChildConstraintsArray(
+  allElementProps: AllElementProps,
+  path: ElementPath,
+): any[] {
+  const value = allElementProps[EP.toString(path)]?.['data-constraints'] ?? []
+  if (!Array.isArray(value)) {
+    return []
+  }
+  return value
+}
+
+function setGroupChildConstraint(
+  dispatch: EditorDispatch,
+  option: GroupChildConstraintOptionType | 'toggle',
+  selectedViews: ElementPath[],
+  allElementProps: AllElementProps,
+  dimension: LayoutPinnedProp,
+) {
+  function makeNewProps(current: any[], mode: 'add' | 'remove'): any[] {
+    switch (mode) {
+      case 'add':
+        return [...current, dimension]
+      case 'remove':
+        return current.filter((other) => other !== dimension)
+      default:
+        assertNever(mode)
+    }
+  }
+  function getMode(): 'add' | 'remove' {
+    if (option !== 'toggle') {
+      return option === 'constrained' ? 'add' : 'remove'
+    }
+    const notAllContainDimension = selectedViews.some(
+      (path) => !getSafeGroupChildConstraintsArray(allElementProps, path).includes(dimension),
+    )
+    return notAllContainDimension ? 'add' : 'remove'
+  }
+  const mode = getMode()
+
+  const prop = PP.create('data-constraints')
+
+  const actions: EditorAction[] = mapDropNulls((path) => {
+    const constraints = getSafeGroupChildConstraintsArray(allElementProps, path)
+    const newProps = makeNewProps(constraints, mode)
+    const uniqueNewProps = uniqBy(newProps, (a, b) => a === b)
+    return uniqueNewProps.length === 0
+      ? unsetProperty(path, prop)
+      : setProp_UNSAFE(path, prop, jsExpressionValue(uniqueNewProps, emptyComments))
+  }, selectedViews)
+
+  dispatch(actions)
+}
