+++ editor/src/components/inspector/sizing-section.tsx
@@ -1,6 +1,5 @@
 import React from 'react'
 import { FlexRow, InspectorSubsectionHeader } from '../../uuiui'
-import { EditorContractDropdown } from './editor-contract-section'
 import { FillHugFixedControl } from './fill-hug-fixed-control'
 import { ResizeToFitControl } from './resize-to-fit-control'
 
@@ -14,9 +13,9 @@ export const SizingSection = React.memo<SizingSectionProps>(() => {
           style={{
             flexGrow: 1,
             height: 42,
-            justifyContent: 'flex-end',
           }}
         >
+          <span style={{ flex: 1 }}>Resizing</span>
           <ResizeToFitControl />
         </FlexRow>
       </InspectorSubsectionHeader>
